refactoring

+++ src/Yahtzee/ScoreTable.h
@@ -30,7 +30,7 @@ public:
 	~ScoreTable() {}
 	ScoreTable(const ScoreTable& other);
 
-	void AssignScoreIfNotAssigned(Scores::ScoreName scoreName, size_t score, bool markAsAssigned = false);
+	void AssignScoreIfNotAssigned(Scores::ScoreName scoreName, unsigned short score, bool markAsAssigned = false);
 	bool HasScore(Scores::ScoreName scoreName) const;
 	unsigned short GetScoreAt(Scores::ScoreName scoreName) const;
 	unsigned short TotalScore() const;
